Store randomness from previous relay block in pallet storage
Because since #288 collator rotation is handled in on_initialize, and we
can only read the relay randomness after the set_validation_data
inherent, which happens after on_initialize.

The solution is to read the randomness on the on_finalize hook of the
previous block, store it in pallet_collator_assignment, and delete it
when reading it.