Move integration tests inside src folder
Turning them into unit tests according to cargo.

This compiles all unit tests into a single binary, instead of having a
separate binary for each file in the tests folder, so compiling tests
should be much faster now. Also it allows integration tests to use
private methods, which is not needed but it may be nice to have in the
future.

https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html