Add script to print biggest weights according to regex (#992)
We recently had a strange issue with a weight proof size being very wrong: #987

This PR adds a script to detect such invalid weights, either in the whole repo or in the PR diff. The PR diff runs in CI on every PR, and if it finds a new big weight it will leave a comment.

Usage: ./tools/show-me-the-weights.py