rewards_per_chain should not depend on chains without collators (#1412)
We have a small bug in pallet_inflation_rewards, where we calculate the reward per chain based on the total number of registered chains, regardless of whether they have collators or not. This case doesn't happen in any live runtime because we have enough collators for all chains (for now).

In the case of some chains without collators, there is a portion of the inflation that always goes to the parachain bond. This PR fixes this, increasing the rewards_per_chain in that case. It doesn't affect inflation, the total inflated value is still the same. (except in the edge case where we have 0 collators, because now we won't mint anything).